Remove some shader-related circular dependencies
