+++ src/deprecated.js
@@ -220,7 +220,7 @@ import {
 } from './graphics/graphics.js';
 import { drawQuadWithShader } from './graphics/simple-post-effect.js';
 import { programlib } from './graphics/program-lib/program-lib.js';
-import { shaderChunks } from './graphics/chunks.js';
+import { shaderChunks } from './graphics/program-lib/chunks/chunks.js';
 import { GraphicsDevice } from './graphics/device.js';
 import { IndexBuffer } from './graphics/index-buffer.js';
 import { createFullscreenQuad, drawFullscreenQuad, PostEffect } from './graphics/post-effect.js';
