+++ src/graphics/program-lib/particle.js
@@ -1,6 +1,7 @@
 import { BLEND_ADDITIVE, BLEND_MULTIPLICATIVE, BLEND_NORMAL } from '../../scene/constants.js';
-import { shaderChunks } from '../chunks.js';
+import { shaderChunks } from './chunks/chunks.js';
 
+import { collectAttribs } from './utils.js';
 import { programlib } from './program-lib.js';
 
 var particle = {
@@ -14,17 +15,14 @@ var particle = {
         return key;
     },
 
-    _animTex: function (options, chunk) {
+    _animTex: function (options) {
         var vshader = "";
-        vshader += options.animTexLoop ? chunk.particleAnimFrameLoopVS : chunk.particleAnimFrameClampVS;
-        vshader += chunk.particleAnimTexVS;
+        vshader += options.animTexLoop ? shaderChunks.particleAnimFrameLoopVS : shaderChunks.particleAnimFrameClampVS;
+        vshader += shaderChunks.particleAnimTexVS;
         return vshader;
     },
 
     createShaderDefinition: function (device, options) {
-
-        var chunk = shaderChunks;
-
         var vshader = "";
         var fshader = programlib.precisionCode(device) + "\n";
 
@@ -44,36 +42,36 @@ var particle = {
         if (options.normal == 1) vshader += "\nvarying vec3 Normal;\n";
         if (options.soft) vshader += "\nvarying float vDepth;\n";
 
-        var faceVS = options.customFace ? chunk.particle_customFaceVS : chunk.particle_billboardVS;
+        var faceVS = options.customFace ? shaderChunks.particle_customFaceVS : shaderChunks.particle_billboardVS;
 
         if (!options.useCpu) {
-            vshader += chunk.particle_initVS;
-            vshader += (options.pack8 ? chunk.particleInputRgba8PS : chunk.particleInputFloatPS);
-            if (options.soft > 0) vshader += chunk.screenDepthPS;
-            vshader += chunk.particleVS;
-            if (options.localSpace) vshader += chunk.particle_localShiftVS;
-            if (options.animTex) vshader += this._animTex(options, chunk);
-            if (options.wrap) vshader += chunk.particle_wrapVS;
-            if (options.alignToMotion) vshader += chunk.particle_pointAlongVS;
-            vshader += options.mesh ? chunk.particle_meshVS : faceVS;
-            if (options.normal == 1) vshader += chunk.particle_normalVS;
-            if (options.normal == 2) vshader += chunk.particle_TBNVS;
-            if (options.stretch > 0.0) vshader += chunk.particle_stretchVS;
-            vshader += chunk.particle_endVS;
-            if (options.soft > 0) vshader += chunk.particle_softVS;
+            vshader += shaderChunks.particle_initVS;
+            vshader += (options.pack8 ? shaderChunks.particleInputRgba8PS : shaderChunks.particleInputFloatPS);
+            if (options.soft > 0) vshader += shaderChunks.screenDepthPS;
+            vshader += shaderChunks.particleVS;
+            if (options.localSpace) vshader += shaderChunks.particle_localShiftVS;
+            if (options.animTex) vshader += this._animTex(options);
+            if (options.wrap) vshader += shaderChunks.particle_wrapVS;
+            if (options.alignToMotion) vshader += shaderChunks.particle_pointAlongVS;
+            vshader += options.mesh ? shaderChunks.particle_meshVS : faceVS;
+            if (options.normal == 1) vshader += shaderChunks.particle_normalVS;
+            if (options.normal == 2) vshader += shaderChunks.particle_TBNVS;
+            if (options.stretch > 0.0) vshader += shaderChunks.particle_stretchVS;
+            vshader += shaderChunks.particle_endVS;
+            if (options.soft > 0) vshader += shaderChunks.particle_softVS;
         } else {
-            if (options.soft > 0) vshader += chunk.screenDepthPS;
-            vshader += chunk.particle_cpuVS;
-            if (options.localSpace) vshader += chunk.particle_localShiftVS;
-            if (options.animTex) vshader += this._animTex(options, chunk);
-            // if (options.wrap) vshader += chunk.particle_wrapVS;
-            if (options.alignToMotion) vshader += chunk.particle_pointAlongVS;
-            vshader += options.mesh ? chunk.particle_meshVS : faceVS;
-            if (options.normal == 1) vshader += chunk.particle_normalVS;
-            if (options.normal == 2) vshader += chunk.particle_TBNVS;
-            if (options.stretch > 0.0) vshader += chunk.particle_stretchVS;
-            vshader += chunk.particle_cpu_endVS;
-            if (options.soft > 0) vshader += chunk.particle_softVS;
+            if (options.soft > 0) vshader += shaderChunks.screenDepthPS;
+            vshader += shaderChunks.particle_cpuVS;
+            if (options.localSpace) vshader += shaderChunks.particle_localShiftVS;
+            if (options.animTex) vshader += this._animTex(options);
+            // if (options.wrap) vshader += shaderChunks.particle_wrapVS;
+            if (options.alignToMotion) vshader += shaderChunks.particle_pointAlongVS;
+            vshader += options.mesh ? shaderChunks.particle_meshVS : faceVS;
+            if (options.normal == 1) vshader += shaderChunks.particle_normalVS;
+            if (options.normal == 2) vshader += shaderChunks.particle_TBNVS;
+            if (options.stretch > 0.0) vshader += shaderChunks.particle_stretchVS;
+            vshader += shaderChunks.particle_cpu_endVS;
+            if (options.soft > 0) vshader += shaderChunks.particle_softVS;
         }
         vshader += "}\n";
 
@@ -92,33 +90,33 @@ var particle = {
         fshader += programlib.tonemapCode(options.toneMap);
 
         if (options.fog === 'linear') {
-            fshader += chunk.fogLinearPS;
+            fshader += shaderChunks.fogLinearPS;
         } else if (options.fog === 'exp') {
-            fshader += chunk.fogExpPS;
+            fshader += shaderChunks.fogExpPS;
         } else if (options.fog === 'exp2') {
-            fshader += chunk.fogExp2PS;
+            fshader += shaderChunks.fogExp2PS;
         } else {
-            fshader += chunk.fogNonePS;
+            fshader += shaderChunks.fogNonePS;
         }
 
         if (options.normal == 2) fshader += "\nuniform sampler2D normalMap;\n";
-        if (options.soft > 0) fshader += chunk.screenDepthPS;
-        fshader += chunk.particlePS;
-        if (options.soft > 0) fshader += chunk.particle_softPS;
+        if (options.soft > 0) fshader += shaderChunks.screenDepthPS;
+        fshader += shaderChunks.particlePS;
+        if (options.soft > 0) fshader += shaderChunks.particle_softPS;
         if (options.normal == 1) fshader += "\nvec3 normal = Normal;\n";
-        if (options.normal == 2) fshader += chunk.particle_normalMapPS;
-        if (options.normal > 0) fshader += options.halflambert ? chunk.particle_halflambertPS : chunk.particle_lambertPS;
-        if (options.normal > 0) fshader += chunk.particle_lightingPS;
+        if (options.normal == 2) fshader += shaderChunks.particle_normalMapPS;
+        if (options.normal > 0) fshader += options.halflambert ? shaderChunks.particle_halflambertPS : shaderChunks.particle_lambertPS;
+        if (options.normal > 0) fshader += shaderChunks.particle_lightingPS;
         if (options.blend == BLEND_NORMAL) {
-            fshader += chunk.particle_blendNormalPS;
+            fshader += shaderChunks.particle_blendNormalPS;
         } else if (options.blend == BLEND_ADDITIVE) {
-            fshader += chunk.particle_blendAddPS;
+            fshader += shaderChunks.particle_blendAddPS;
         } else if (options.blend == BLEND_MULTIPLICATIVE) {
-            fshader += chunk.particle_blendMultiplyPS;
+            fshader += shaderChunks.particle_blendMultiplyPS;
         }
-        fshader += chunk.particle_endPS;
+        fshader += shaderChunks.particle_endPS;
 
-        var attributes = shaderChunks.collectAttribs(vshader);
+        var attributes = collectAttribs(vshader);
 
         return {
             attributes: attributes,
